+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [3.1.2]
+- Refactor and improve StructureConfig class
+- Refactor ConfigManager class
+
 ## [3.1.1]
 - Band-Aid fix for issues with the core config system